api/requirements.txt
fastapi
uvicorn
pandas<2.2
psycopg2-binary